requirements.txt
pandas
numpy
scikit-learn
faiss-cpu
gradio
transformers
torch
